# backend/docs/swagger.yaml
basePath: /
definitions:
  api.adminConfigResponse:
    properties:
      registrationEnabled:
        type: boolean
    type: object
  api.adminEventDTO:
    properties:
      allowEdit:
        type: boolean
      configShowRanking:
        type: boolean
      createdAt:
        type: string
      id:
        type: string
      interactionsEnabled:
        type: boolean
      ownerEmail:
        type: string
      ownerId:
        type: string
      ownerName:
        type: string
      participantCount:
        type: integer
      pinCode:
        type: string
      questionCount:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
  api.adminEventsResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/api.adminEventDTO'
        type: array
    type: object
  api.adminResetLinkResponse:
    properties:
      expiresAt:
        description: ExpiresAt no formato RFC3339.
        type: string
      token:
        type: string
    type: object
  api.adminUserDTO:
    properties:
      createdAt:
        type: string
      email:
        type: string
      eventCount:
        type: integer
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  api.adminUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/api.adminUserDTO'
        type: array
    type: object
...
  api.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.okResponse:
    properties:
      ok:
        type: boolean
    type: object
...
  api.redefinirSenhaRequest:
    properties:
      novaSenha:
        type: string
      token:
        type: string
    type: object
...
  api.updateAdminConfigRequest:
    properties:
      registrationEnabled:
        type: boolean
    type: object
...
  api.validTokenResponse:
    properties:
      valid:
        type: boolean
    type: object
info:
  contact: {}
  description: |-
    API REST do Arandu (coleta de respostas, apresentação ao vivo e administração de eventos) — as mesmas rotas usadas pelo frontend embutido, prontas para chamar diretamente e criar/gerenciar recursos por API.

    Autenticação de organizador (cookieAuth): faça login em outra aba pelo próprio app (POST /api/conta/entrar) — o cookie de sessão é httpOnly, então não dá para colá-lo aqui manualmente, mas o navegador o envia sozinho nas chamadas desta página por estarem na mesma origem.
    Visitante ao vivo (liveViewerToken): obtenha o token em POST /api/publico/eventos/{id}/ao-vivo/entrar e cole em "Authorize" — o Swagger UI passa a anexar "?token=" nas chamadas das rotas públicas de ao-vivo.
  license:
    name: Uso interno
  title: Arandu API
  version: "1.0"
paths:
  /api/admin/configuracoes:
    get:
      description: Somente super admin. Hoje só tem o toggle de cadastro aberto.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.adminConfigResponse'
        "403":
          description: acesso restrito ao super admin
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - cookieAuth: []
      summary: Configurações administrativas
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: Somente super admin. Liga/desliga o cadastro de novas contas organizadoras.
      parameters:
      - description: Nova configuração
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateAdminConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.adminConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: acesso restrito ao super admin
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - cookieAuth: []
      summary: Atualiza configurações administrativas
      tags:
      - admin
  /api/admin/eventos:
    get:
      description: Somente super admin.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.adminEventsResponse'
        "403":
          description: acesso restrito ao super admin
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - cookieAuth: []
      summary: Lista todos os eventos, de todos os organizadores
      tags:
      - admin
  /api/admin/usuarios:
    get:
      description: Somente super admin — cada usuário vem com sua contagem de eventos.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.adminUsersResponse'
        "403":
          description: acesso restrito ao super admin
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - cookieAuth: []
      summary: Lista todos os usuários
      tags:
      - admin
  /api/admin/usuarios/{id}:
    delete:
      description: Somente super admin. Quem está agindo não pode se auto-excluir
        por aqui.
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.okResponse'
        "400":
          description: auto-exclusão bloqueada
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: acesso restrito ao super admin
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - cookieAuth: []
      summary: Exclui um usuário (e seus eventos)
      tags:
      - admin
  /api/admin/usuarios/{id}/link-redefinicao:
    post:
      description: Somente super admin. Token de uso único, válido por 1h — o admin
        copia e envia por fora (o app não envia nada); gerar de novo invalida o anterior.
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.adminResetLinkResponse'
        "403":
          description: acesso restrito ao super admin
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      security:
      - cookieAuth: []
      summary: Gera um link de redefinição de senha para um usuário
      tags:
      - admin
...
  /api/publico/redefinir-senha:
    get:
      description: Público, sem sessão. Usado por /redefinir-senha antes de mostrar
        o formulário de nova senha.
      parameters:
      - description: Token do link gerado pelo super admin
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.validTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "404":
          description: link inválido ou expirado
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Valida um link de redefinição de senha
      tags:
      - conta
    post:
      consumes:
      - application/json
      description: Público, sem sessão. O token é de uso único e expira em 1h.
      parameters:
      - description: Token e nova senha
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.redefinirSenhaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.okResponse'
        "400":
          description: link inválido/expirado ou senha fora das regras
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Define uma nova senha a partir do link de redefinição
      tags:
      - conta
securityDefinitions:
  cookieAuth:
    in: header
    name: Cookie
    type: apiKey
  liveViewerToken:
    in: query
    name: token
    type: apiKey
swagger: "2.0"